:building_construction: #12 refactor module construction
